fix(tanstack-react): add TanStack Query integration with new adapter functions
- Introduced `asTanStackEndpoint` and `createTanStackAPI` functions to enhance Fezi endpoints with TanStack Query capabilities.
- Updated `.prettierignore` to exclude TypeScript declaration files except for `vite.config.d.ts`.
- Refactored `index.ts` to export the new adapter functions.
