+++ .prettierignore
@@ -2,3 +2,4 @@ dist
 node_modules
 coverage
 *.d.ts
+!vite.config.d.ts
